第二版本待修改

+++ src/main/java/DAO/ViewrecordDAOImpl.java
@@ -14,13 +14,14 @@ public class ViewrecordDAOImpl implements ViewrecordDAO {
     Connection conn = dbc.getConnection();
 
     public Boolean insert(Viewrecord vr) {
-        String sql = "insert into viewrecord values (?,?,?)";
+        String sql = "insert into viewrecord values (?,?,?,?)";
         PreparedStatement ps = null;
         try {
             ps = conn.prepareStatement(sql);
-            ps.setString(1,vr.getCar_id());
-            ps.setString(2,vr.getCususer_id());
-            ps.setTimestamp(3,vr.getView_date());
+            ps.setString(1,vr.getShopuser_id());
+            ps.setString(2,vr.getCar_id());
+            ps.setString(3,vr.getCususer_id());
+            ps.setTimestamp(4,vr.getView_time());
 
             ps.executeUpdate();
             ps.close();
@@ -32,13 +33,14 @@ public class ViewrecordDAOImpl implements ViewrecordDAO {
     }
 
     public Boolean update(Viewrecord vr){
-        String sql = "update viewrecord set car_id=?,cususer_id=?,view_date=?";
+        String sql = "update viewrecord set shopuser_id=?,car_id=?,cususer_id=?,view_date=?";
         PreparedStatement ps = null;
         try {
             ps = conn.prepareStatement(sql);
-            ps.setString(1,vr.getCar_id());
-            ps.setString(2,vr.getCususer_id());
-            ps.setTimestamp(3,vr.getView_date());
+            ps.setString(1,vr.getShopuser_id());
+            ps.setString(2,vr.getCar_id());
+            ps.setString(3,vr.getCususer_id());
+            ps.setTimestamp(4,vr.getView_time());
 
             ps.executeUpdate();
             ps.close();
@@ -49,14 +51,15 @@ public class ViewrecordDAOImpl implements ViewrecordDAO {
         }
     }
 
-    public Boolean delete(String car_id, String cususer_id,Timestamp view_time){
-        String sql = "delete from viewrecord where car_id=?,cususer_id=?,view_time=?";
+    public Boolean delete(String shopuser_id, String car_id, String cususer_id,Timestamp view_time){
+        String sql = "delete from viewrecord where shopuser_id=? and car_id=? and cususer_id=? and view_time=?";
         PreparedStatement ps = null;
         try {
             ps = conn.prepareStatement(sql);
-            ps.setString(1,car_id);
-            ps.setString(2,cususer_id);
-            ps.setTimestamp(3,view_time);
+            ps.setString(1,shopuser_id);
+            ps.setString(2,car_id);
+            ps.setString(3,cususer_id);
+            ps.setTimestamp(4,view_time);
 
             ps.executeUpdate();
             ps.close();
@@ -67,21 +70,22 @@ public class ViewrecordDAOImpl implements ViewrecordDAO {
         }
     }
 
-    public Viewrecord findById(String car_id, String cususer_id,Timestamp view_time){
+    public Viewrecord findById(String shopuser_id, String car_id, String cususer_id,Timestamp view_time){
         Viewrecord vr = new Viewrecord();
-        String sql = "select * from viewrecord where car_id=?,cususer_id=?,view_time=?";
+        String sql = "select * from viewrecord where shopuser_id=? and car_id=? and cususer_id=? and view_time=?";
         PreparedStatement ps = null;
         ResultSet rs = null;
         try {
             ps = conn.prepareStatement(sql);
-            ps.setString(1,car_id);
-            ps.setString(2,cususer_id);
-            ps.setTimestamp(3,view_time);
+            ps.setString(1,shopuser_id);
+            ps.setString(2,car_id);
+            ps.setString(3,cususer_id);
+            ps.setTimestamp(4,view_time);
             rs = ps.executeQuery();
             if(rs.next()){
                 vr.setCar_id(rs.getString(1));
                 vr.setCususer_id(rs.getString(2));
-                vr.setView_date(rs.getTimestamp(3));
+                vr.setView_time(rs.getTimestamp(3));
             }
 
 
@@ -101,9 +105,10 @@ public class ViewrecordDAOImpl implements ViewrecordDAO {
             rs = ps.executeQuery();
             while(rs.next()){
                 Viewrecord vr = new Viewrecord();
-                vr.setCar_id(rs.getString(1));
-                vr.setCususer_id(rs.getString(2));
-                vr.setView_date(rs.getTimestamp(3));
+                vr.setShopuser_id(rs.getString(1));
+                vr.setCar_id(rs.getString(2));
+                vr.setCususer_id(rs.getString(3));
+                vr.setView_time(rs.getTimestamp(4));
 
                 vrList.add(vr);
             }
